Move NFT components to its own folder
