+++ src/renderer/components/Nft/Image.js
