+++ src/renderer/components/Nft/Skeleton.js
