+++ src/renderer/drawers/NFTViewerDrawer/index.js
@@ -22,8 +22,8 @@ import { NFTProperties } from "./NFTProperties";
 import { CopiableField } from "./CopiableField";
 import NftPanAndZoom from "./NftPanAndZoom";
 import ExternalViewerButton from "./ExternalViewerButton";
-import Skeleton from "~/renderer/screens/nft/Skeleton";
-import Image from "~/renderer/screens/nft/Image";
+import Skeleton from "~/renderer/components/nft/Skeleton";
+import Image from "~/renderer/components/nft/Image";
 import { openModal } from "~/renderer/actions/modals";
 import { setDrawer } from "~/renderer/drawers/Provider";
 import { SplitAddress } from "~/renderer/components/OperationsList/AddressCell";
