+++ src/renderer/screens/nft/Gallery/TokensList/Item.js
@@ -10,8 +10,8 @@ import { getNFTById } from "~/renderer/reducers/accounts";
 import Box, { Card } from "~/renderer/components/Box";
 import Text from "~/renderer/components/Text";
 import { centerEllipsis } from "~/renderer/styles/helpers";
-import Image from "~/renderer/screens/nft/Image";
-import Skeleton from "~/renderer/screens/nft/Skeleton";
+import Image from "~/renderer/components/nft/Image";
+import Skeleton from "~/renderer/components/nft/Skeleton";
 import IconDots from "~/renderer/icons/Dots";
 import { useNftMetadata } from "@ledgerhq/live-common/lib/nft";
 import NFTContextMenu from "~/renderer/components/ContextMenu/NFTContextMenu";
